Demonstrate one way to create Rust errors and integrate with Python
While it's possible to create your error types by hand, I like to use
my library [SNAFU][snafu] to help with some of the boilerplate. There
are other popular libraries such as `thiserror` or `anyhow` as well.

I demonstrate a few facets of the crate here:

1. Creating opaque errors that do not expose internal details. Here,
   `ir_to_dimble::Error` wraps `ir_to_dimble::InnerError` and hides
   details except for the causal chain and error text. This is good
   for public APIs where you should be mindful of semver concerns.

2. Localized errors that compose into larger
   ones. `ir_to_dimble::SerialiseFieldsError` only deals with the ways
   that `serialise_dimble_fields` can fail and helps you avoid
   repetition and keep the scope a little tighter.

I also followed PyO3's example and implemented a conversion from the
Rust error to a `PyErr`, going through the existing
`DimbleError`. This is a very basic implementation; a better one would
expose the causal chain as a Python exception chain and potentially
even have a traceback including Rust code.

[snafu]: https://docs.rs/snafu/